[python] Move LeRobot import API out of package init
